added powerpoint and modified readme
