--- .gitignore
@@ -1,3 +1,0 @@
-*.xlsx
-*.txt
-*.csv
