--- delete_list.csv
@@ -1,18 +1,0 @@
-Delete List
-http://www.hotstuffracing.co.za
-https://www.avforums.co.za
-https://vwclub.co.za
-https://www.aircooledvwsa.co.za
-http://www.4x4community.co.za
-http://www.approvedauto.co.za
-https://www.dispatchlive.co.za
-https://eazyaccess.co.za
-http://digitalcatalogs.net
-http://www.circle7cars.co.za
-http://www.wilddog.za
-https://www.imotoonline.co.za
-https://www.publicads.co.za
-https://aellaaellaaella.wordpress.com
-https://mobserver.co.za/83016
-http://www.kznmotors.co.za
-https://www.classads.co.za
