--- url database.txt
@@ -1,34 +1,0 @@
-http://www.cosmopolitan.co.za - 403170
-http://glamour.co.za - 59870
-http://thedailyvox.co.za - 131740
-https://sagamer.co.za - 272070
-https://timeslive.co.za - 13140000
-https://news365.co.za - 264100
-http://www.gq.co.za - 147040
-https://sowetanlive.co.za - 5900000
-http://allafrica.com - 2530000
-http://sabcnews.com - 541250
-http://ewn.co.za - 13170000
-https://enca.co.za - 8580000
-http://nag.co.za  - 100000
-http://947.co.za - 245830
-https://techcentral.co.za - 299310
-http://glamour.co.za - 59870
-http://702.co.za - 769250
-http://kfm.co.za - 197490
-http://capetalk.co.za - 365410
-http://www.popularmechanics.co.za - 50000
-https://dstv.com - 10480000
-https://iol.co.za - 15700000
-https://mweb.co.za - 3040000
-https://mybroadband.co.za - 11150000
-https://arrivealive.co.za - 228380
-https://news24.com - 45610000
-https://peoplemagazine.co.za - 224310
-https://mg.co.za - 3060000
-http://transafricaradio.net - 99940
-http://mzansiville.co.za - 90440
-http://702.co.za - 769250
-http://947.co.za - 245830
-https://netwerk24.com - 2900000
-http://transafricaradio.net - 99940
